core-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
google-auth==2.35.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
eth-account==0.13.7
cryptography==43.0.1
httpx==0.27.2
requests==2.33.1
